build.sh: pass args for hsh via ../apt/test-pk-in-hsh.sh, too
(Like --apt-conf etc.)

The script ../apt/test-pk-in-hsh.sh has been prepared for this
since this commit in apt Git repo:

d831cfd4d test-pk-in-hsh.sh: allow to pass more opts for hsh --ini

BTW, a convenient thing about this is:

the passed --apt-conf option will not only affect the sources.list
for the hasher, but also (with --ini SAME) the sources.list set up inside
the hasher for apt (via hsh-setup-apt-with-same-sources-via-ssh "$HSHDIR").
The hasher "stuff" won't be seen by apt inside the hasher, however,
because it is normally of "rpm-dir" type, which doesn't work non-locally,
i.e., although the corresponding URL might be present in sources.list,
the content (the packages themselves) won't be seen.